refactor: organize auth page workflow
