+++ frontend/src/pages/auth/components/AuthErrorBanner.tsx
