+++ frontend/src/pages/auth/components/AuthSignupNameFields.tsx
@@ -1,7 +1,7 @@
 import { AnimatePresence, motion } from 'motion/react'
 import { AuthTextField } from './AuthTextField'
-import { SIGNUP_FIELD_ANIMATION } from './authAnimations'
-import type { AuthFieldErrors, AuthFormValues } from './authForm'
+import { SIGNUP_FIELD_ANIMATION } from '@/pages/auth/constants/authAnimations'
+import type { AuthFieldErrors, AuthFormValues } from '@/pages/auth/utils/authForm'
 
 interface AuthSignupNameFieldsProps {
   errors: AuthFieldErrors
@@ -51,4 +51,3 @@ export function AuthSignupNameFields({
     </AnimatePresence>
   )
 }
-
