+++ frontend/src/pages/auth/constants/authAnimations.ts
